requirements.txt
flask
azure-cosmos
gunicorn